Rename GamePage to FillInBlankGamePage, rename SlideBox to AnswerChoiceBox
